Added rounding to read_TPS
The different precisions in the RPE and TPS data have been causing
problems, for variables like F_4, where only few unique values exist.
Rounding the TPS data to the same precision as the RPE data prevents it
from distorting the pB values for the regions between the sparse RPE
values, which would yield a vast overabundance of pB = 1 grid points.