Fix x and y values used for prediction calc_partial_map_generated
Prior x and y values used for the prediction of the value in each cell
were the integers obtained from the binning. These very high inout values
caused very high predictions.

Adding a np.linspace going from the minimal to the maximal value and using
the x and y values as indices of this linspace fixes the issue.