Refactored plot_super_map() and added calc_partial_map_given()
Changed plot_super_map() to take in the method to be used for each subplot
as an argument instead of determining this via if/else statements.
Required a change of input variables and an identical set of input variables
for all possible methods that can be passed in.

Added calc_partial_map_given() to only plot points of interest for the
given snapshots and make it comparable with the partial map of predicted
points. For the partial map, the full map is calculated and then only
relevant points are retained.