Added trimmer.py
Added file with a Trimmer class. The object is instantiated with an ordered
list of pBs. self._keep_elements is generated as a boolean mask based on where
pB values are either 0 or 1.
trim() uses the mask to remove all elements of a provided list that conincide
with these pB values of 0 or 1.