Change balance_and_trim() to balance_and_mask() and add trim_and_balance()
As balance_and_trim() did not actually trim the values but just masked
them to 0, it was renamed to balance_and_mask() and the variable names
changed accordingly.
A new static method trim_and_balance() was added that first trims the pBs
by removing all entries of 0 and 1, and then calculates the balancing weights.