Replace dataO.minima and dataO.maxima with reduced version
dataO.minima and dataO.maxima had not been reduced yet to only contain
the variables that are being used. When generating a map, the area for
which predictions were made was spanning a from different minima and maxima
than intended. By generating train_reduced_minima and train_reduced_maxima
and using them for plotting, this problem is solved.