Allow for extensive testing of different datasets and conditions
Add a for loop to iterate over different methods to be tested with the
different datasets.

As full_pB_balanced_weights in datasets seem to not yield any good results
their test section is removed.
Same holds true for the standard labels and standard weights dataset
and pBs as weights datasets.