Rename variables and remove unneeded elements of initialization
To keep apart the snapshots that are used for initialization and
the snapshots that are taken as an argument of bound_snapshots() the former
is renamed to base_snapshots.
Storing of self._snapshots and self._outlier_cutoff is omitted as both
variables are not needed outside of the __init__ function.