Wrote class that constitutes the pipeline for data processing.
As all used data will be processed in the same way, it makes sense to make a
pipeline. When initializing the pipeline a list of base_snapshots is taken as
an argument which forms the basis of all processing steps (e.g. bounding,
normalization,...). The processing classes are initialized in order and stored
in self.

Any data that will be used can be processed via the process() function. Here
the processing classes are consecutively applied to the data.

The order in which the processing classes are aplied can be changed easily,
but this should be done in both the __init__ dunder and in process().