Added three new processing functions rbnga(), rbngat() and rbngatb()
rbnga() acts as rbng() but additionally  approximates the pBs based
on the processed snapshots as well as given labels and weights.

rbngat() acts as rbnga() but additionally uses the approximated pBs to
trim the snapshots, labels, weights, pB_dict and pBs.

rbngatb() acts as rbgnat() but additionally generates balanced weights for
the pBs.