Refactor DatasetData such that each instance only contains train, val or test data
Change the __init__() of DatasetData to oly take in the data of one dataset
type instead of all three and add a flag to indicate which tiype of data it
contains.
Adapt the string method, to print the flag and size of the dataset.
Remove outdated properties used for the different dataset types.
Write a static method to initialize all three dataset types at once with
data coming from SnapData.

Reoder the output from SnapData's split_lists() to fit to the DatasetData's
initialize_train_val_test_datasets() function.