Adapt pipeline to new form of DatasetData and fix labels variable name
Change rbnga(), rbngat() and rbngatb() to take in DatasetData objects as
these now only contain relevant data anyways and passing of several variables
can therefore be avoided.
Change plot_data(), importanct_data() and stepwise_data() for the same reasons
but ask for specific dataset types. If the flag is incorrect and
AssertionError is thrown.

Rename label to labels.