Pass GlobalConstants object into PlotData and ad logvmin variable
To reduce the number of input variables, pass a GlobalCostants object into
the functions of PlotData. Remoe resolution, vmin, vmax and subfigsize from
the argmuments for plot_super_map() and/or plot_super_scatter()

Add logvmin variable to the global constant to make it possible to adapt
the lower end of a logarithmic colorbar. Add the corresponding variable in
the PlotData functions.