Make a file for helper functions and add get_size() and make_halfpoint_divided_colormap()
Move get_size() from NPA_v8 to the helper function file.
Write a new function to generate a colormap based on "summer" but with a small
range around 0.5 at which the color is black. This allows a visualization
of the commitor surface.
As argmunet the function takes logvmin, to adapt to different lower bounds
for the colormap.