Modify make_halfpoint_divided_colormap() to have the same bandwidth with different logvmin
Define a bandwidth and calculate the lower and upper bound accordingly on
the log scale.
Use the upper and lower bounds, as well as their absolut difference to
define the cmap.
After some testing set bandwidth to 0.3 to get a well visible 0.5 dividing
surface.
Increase resolution to 1001 to be less prone to rounding errors for different
logvmin values.