Refactor plotData.py to make functions staticmethods, rename to plotter.py
Rename PlotData class to Plotter and plotData.py to plotter.py.
Make all functions of Plotter static, by removing references to self and
passing in all needed arguments instead. This avoids having to initialize
a Plotter object with data needed for e.g. calc_map_given() when only
calc_map_generated() will be used.
Remove the initialization from the Plotter class, as it is not needed anymore.
Remove outdated parts of the code in plotter.py
In globalConstants.py split the stamp property into data_stamp and model_stamp
properties as the full information is not always needed.
Adapt the naming of plots generated in plotter.py to only contain the stamps
that are relevant.
Remove the plot_data() function from pipeline.py as this was used for
initilization of PlotData/Plotter objects, which is now outdated.