Attempt fixing the difference between given map made from labels and weights vs. pBs
Adapt pB_Approximator to also return a list of pB_weights. These contain the
summed weights used for the pB_approximation of a given point. Not having
these meant so far, that when pBs were binned from more dimensions, pBs
made from many low weight points would count as much as pBs made from many
high weight points. This biased the results of the 2D-Projection.
Adapt Pipeline so that the functions also return the pB_weights while the
original variable pB_weights is renamed to pBb_weights (as those are the
balanced weigths).
Add sections to NPA to generate super_maps with different inputs.
Write compare_super_maps() to see which settings yield the same super_map.
However, for the trimmed training data, none of the super_maps are the same.
With the untrimmed shooting points, standard labels and standard weights are
very similar to pBs and pB_weights. The remaining differinces could be due to
rounding errors.
Trimming seems to have a very negative effect on the 2D respresentation, as
cells with either pB 0 or 1 are not contributing although taken together
they would have yielded a pB between 0 and 1.