Make plotting of ground truth dependent on standard labels and weights and move code to outsourced.py
As it has been identified, that first approximating pBs
and then condensing them onto two dimensions for plotting
might bias the representation in the plot, the only
accurate ground truth is calculated from the labels and
the weights. plot_ground_truth() is adapted to that purpose
and is tailored to only make generated plots based on labels
and weights. The calc_map_generated() calls are moved out of
other functions in NPA as they will not be necessary (or
even misleading) for plotting anything but the ground truth
from standard labels and weights.

Make new file outsourced.py to store currently unused
functions that might be used or checked again later.
Move functions to compare different super_maps there.
Move functions for using the shooting data and plotting
the dividing surface there.
Move miscellaneous functions there.

Clean NPA.ipynb and remove outdated code.