Replace i in range(len()) with i, x in enumerate()
As using enumerate is the more pythonic way of getting the indices of
a list, it replaces all occurences of range(len()) in the code.
At places where both the index and the list element are nedded, both are
being stored. At the other places only the index is kept.