Add and adapt code to test different conditions for the dividing surface
As the dividing surface calculated based on the shootings points does not
match exactly with the pB = 0.5 barrier in the complete dataset different
conditions are being tested.
Using 70% of the dataset
Using only the lower or higher interfaces
Testing also the dividing surface of N_{w,3} vs BigCage

Parameters in globalConstants.py are adapted accordingly and plot_one_map()
and its helper functions are refactored as well.