Change the colorbar used to not display values of 0 and 1 as white.
As only data from points where both AA and AB paths exist should be considered
and points with pB = 0 or pB = 1 mean there are only AA or only AB paths,
these points are not being plotted.
Set the colormap over and under attributes to display white when the end of
the colormap is reached.
Extend the plotted colorbar on both sides to indicate this fact to the viewer.
Change the upper limit of the colorbar from 1 to 1 - const.logvmin, such that
on both sides an equally small fraction of the range is disregarded as 0 or 1.