Change outlier_cutoff from 0.01 to 0.005
The outlier_cutoff influences the projection while plotting but more
importantly can influence the approximated pBs. If a too high cutoff is chosen
values will be capped to points in a region with both AA and AB paths, leading
to an overabundance of AB paths, which would raise the pB value at that point.
Therefore the cutoff should be chosen such that it lies in a region with
only AB oaths to begin with. Capping other AB paths to these points will not
change the approximated pB. 0.005 was found to be a good value for this.
While even lower values can be chosen, a wider range of input values means
that the range of values with prediction power shrinks in comparison.