Add axis labels to the plots from plot_super_map()
Axis labels are based on the upper and lower bound of the dataset,
therefore these two parameters also need to be passsed in.
Similar to axis titles, axis labels are only displayed at the rightmost
and bottom most plots and apply to all the plots in the respective row or
column.