Add renomalization step after the trimming to the pipeline
As the trimming shifts the mean and std of the trimmed_snapshots, they are
no longer normalized.
To renormalize them, the function rbngatn() is added, which takes in the
trimmed snapshots and normalizes them.
rbngatb() is renamed to rbngatnb() as the normalization step takes place
before the balancing.
In NPA the function calls to rbngatb() are renamed correspondingly.
As the minima and maxima of the renormalized, trimmed snapshots do not
match the minima and maxima stored in the pipeline,
plot_with_different_settings() is changed to take in minima and maxima as
agruments.