Make 0 points plotable even with logscale
As log(0) is not defined, points with a value of 0 were not being plotted
in the heatmaps with logarithmic colorbars. Therefore, before plotting values
are compared with the chosen logvmin value / 2 and the maximum is chosen. This
ensures that 0 values are in the color chosed for "under" in the colormap,
while all other points remain unchanged.

To make the visual distinction between values at 0 and 1 possible, the over
and under values of the cmap were changed from grey to a green color for
"under" and an orange color for "over".