Add function to which fraction of the range covered by AB paths is also covered by AA paths
As no BB and BA paths are available it is not clear whether these would
influence the position of the barrier.
If it could be assumed that BB and AA paths show a similar behavior around
state B and A respectively, knowing that how far AA paths reach might be an
indication on how far BB paths might go. If it appears as if BB paths would
not reach beyond the barrier, they will be unlikely to influence the barriers
position.