Extend plot_with_different_settings() to plot and store the progression of the different losses over the epochs
Write plot_loss_history() in helperfuntions.py to make a line plot of the
loss development over the course of the training.
In plot_with_different_settings() save the log history to a TXT file as well
as call plot_loss_history() to make and save the plot of the log history.