Add Hypercube_Balancer class to balance the weights for snapshot reconstruction
Similar to the pBs, the distribution of points along the hypercube is not
uniform. If all points are weighted equally for the reconstruction, the model
will mainly predict lower values as these are vastly more abundant in the
input than the higher values.
To solve this problem, balanced weights are being generated. As with the
pB_Balancer, points occupying similar regions (here hypercubes) of the
parameter space are binned together. Then all bins are assigned the same total
weights that is shared between all points within the respective bin.
This way points from sparse regions will have more weight and all regions
should be weighted equally well.