Fix total weight in pB_Balancer to a constant number
As, so far no weight normalization according to the size of counter took place,
the sum of all weights varied depending on the chosen bin size and the size
of the counter. Like in the Hybercube_Balancer weights are now also divided
by the number of non-empty bins. Thereby it is guaranteed that the total sum
of all weight always equals to the number of snapshots.