Add measure to ensure that the training set consist of equal fractions from each interface
As the different interfaces, carry different information a training set
representing each interface well is important. To ensure that the number
of snapshots originating from a given interface within the training set is
compared with the total number of snapshots from that interface. This value
should as closely as possible resemble the value chosen for the train ratio.

Some of the deeper methods already extracted the origin under the variable
name "name". To avoid ambiguity this was renamed to "origin".
Origin were passed up through the different function calls to
make_train_val_test_split_snapshots_from_snapshots(). Here the number of
snapshots for each interface were counted.
After shuffling and splitting the data set the number of snapshots from each
interface were counted in the training set.
The ratio per interface is printed to the console.