Write MultiDim_Balancer class to balance snapshots according to their distibution along all dimensions
The Hypercube_Balancer did not fully fix the problem of bad reconstruction
for high values. While it ensures that each hypercube has the same total
weight, per dimensions there can still be a non-uniform distribution of
hypercubes along the value range.
In a new approach, each dimension will be balanced separatedly to constitute
the total multi-dimensionally balanced weight.
A similar balancing procedure as in pB_Balancer and Hypercube_Balancer is
used, however here however once per dimension.
The final list of weights is the normalized product of weights for each
dimension.