Improve settings for reconstruction
The fact that reconstructions do not surpass a certain point seems to partly
be due to the fact that the decoder_2 with a tanh actication function can
only output numbers between -1 and 1. Therefore higher numbers cannot be
represented (full range of normalized numbers lies around [-3,7]).
Removing the output layer activation function allows for a broader range
of numbers to be output.
In further testing, the usage of a regularizer of 0.00001, the usage of an
unregularized output layer for decoder_2 as well as hypercube_balanced weights
have been found to give decent results for the reconstruction.
The range is still not fully covered, however, and further improvements
could be attempted.