Add means to set pB = 0 or 1 to higher/lower values instead of trimming
The Trimmer removes all entries with pB = 0 or pB = 1. However there is still
many remaining points with pB close to zero while only few such points exist
close to pB = 1. The predictions of the model seem to reflect that as they
seem to capture the region close to 0 better than the region close to 1.
To compensate for this the Squeezer class was written. Values of pB = 0
are now replaced with the value chosen for logvmin, whie pB = 1 is set to
(1 - logvmin), no trimming occurs.
Within pipeline.py rbngas() and rbngasb() are written as alternative route
to the trimming. Here after approximating the pBs, the squeezing and a
subsequent balancing of pB weights takes place.