Correct mode calculation and add functions to get correlated_representations for all dimensions (pairs)
Fix error in get_modes_from_tuples() as it returns a list of lists, while
a simple list is needed.

Write get_all_1D_correlated_representations() and
get_all_2D_correlated_representations(). These return a dictionary with
the dimension (pair) as key and the representation as value.
This can then be used for a more representative plotting strategy.
Calculate the 1D and 2D representations in NPA.