Add plotting function for corrected scatter plots of reconstruction
Similar to the 2D plots, the 1D scatter plots visualizing the reconstruction
were biased by using 0 as a fill value for all other dimensions.
A function to calculate the 1D_corrected_representations was already written
previously, and now a new method was added to Plotter to make use of that.
Consequently plot_super_scatter() is revised to take in methods as parameter.
Calls to plot_super_scatter() are added to NPA.