Write function to map paths on the latent space
The encoder maps a snapshot onto a latent space consisting of as many
dimensions as there are bottleneck nodes. Low dimensional latent spaces can
be visualized in plots.
map_path_on_2D_latent_space() was added to plotHelperFunctions for this
purpose. In NPA a repeated call of this function was added to generate
plots of different paths onto the latent space.