Add functions to read one path from TIS or TPS data  for latent space plotting
Mapping a complete path onto the latent space of the bottleneck node(s)
allows to see whether a clear trend is apparent. This indicates a RC is
captured.
Instead of reading the complete TPS or TIS data, a function was added to
select single paths.
Thee can then be plotted onto the latent space.
In NPA one call to get_one_TPS_path() and one call per interface to
get_one_TIS_path() were added to examine the paths.