Refactor AutoEncoder.model to reduce duplicate code and allow decoders with 0 hidden layers
model() was renamed to make_models() and refactored. The functions
make_encoder(), make_decoder() and make_autoencoder() were extracted from
it to reduce the amount of duplicate code.
make_encoder() and make_decoder() were additionally rewritten to also allow
0 hidden layers, while the previous lower limit was 1.
Make a separation between the hidden_activation_function for the hidden layers
and the out_put_activation_function for the output layer, which is relevant
for the decoder_2.
Update NPA and plotHelperFuncions.py accordingly.