Change pipeline to return minima and maxima instead of complete list of snapshots for prediction plotting
So far the prepare_prediction_plotter() function returned the complete
bnrn_snapshots, but their only use was to obtain the minima and maxima for
plotting. The function was changed as such, that now minima and maxima
are being calculated withing the function and returned already, to save
memory and time instead of repeatedly computing the minima and maxima.