Add function for generating radial plots of paths reconstructed from the latent space
As the reconstruction_decoder can reconstruct a snapshot from latent space,
it should be possible to reconstruct a path from a set of values uniformly
distributed between the values representing state A in the latent space and
the values representing state B in the latent space.
To facillitate this, plot_example_paths_on_latent_space() was adapted to
return the minimal and maximal value on the latent space it encountered, to be
used as lower and upper limit.
plot_reconstruction_from_latent_space() was added to plotHelperFunctions.py
to generate a radial plot within the bounds found within
plot_example_paths_on_latent_space().
A call to generate the radial plot is added to NPA.