Adapt plot_example_paths_on_latent_space() to let the latent paths span the same range on the y-axis of the plots.
Before every xth snapshot of a path was used to generate and plot the
latent paths. This however, made short paths barely perceptible when plotted
next to longer paths.
make_latent_path_from_path() was adapted to not take every xth snapshots, but
to take a predefined number of equidistantly indexed snapshots. This ensures
that all paths are plotted to be of equal lenght.
Consequntly the y-axis was changed to "Fraction of path length"
The plot legend was adapted to display the Interace names in math mode.
NPA was changeto match the function interfaces.