Make plot_example_paths_on_latent_space() dependent on dataSetType variable
To reduce duplicate code the dataSetType variable is used to select the
get_paths_function. Using that plot_example_paths_on_latent_space() will
extract the paths corresponding to the current dataset, abolishing the need
to manually change parameters.
In plotHelperFunctions.py get_toy_paths() is changed to not take folder_name
as an argument anymore, as the folder name is now also contained in the const
object the function receives.