Change Plotter class to a collection of functions
As the Plotter only constisted of staticmethods and required no initialization,
having to access all its functions via calls to Plotter.* seemed impractical.
Therefore the class was dissolved such that plotter.py is now a collection of
functions that can be imported.
plotHelperfunctions.py and NPA.ipynb were changed accordingly.