Add the choice of loss function for autoencoder_1 and _2 to globalConstants.py
For testing purposes the loss functions were freely selectable and changable.
However, with several functions where the loss functions needs to be
specified as a parameter, the updating all function calls has become cumbersome.
Therefore the Const class was extended by the variables loss_function_1 and
loss_function_2.
All functions compiling models, were changed to not take the loss functions
as an argument but to take it from the Const object they are handed anyways.