Allow for another plot function to be injected into plot_single_map() and add injection of PES overlay.
Sometimes single maps without any additional elements are needed by sometimes
other figures are wanted. Therefore plot_single_map() was adapted to take
in another function that is then additionally executed before the heat map
is generated. lambda x: None is set as the default value for the function
to continue running if no function is passed.
As a first function, inject_PES() is added, which adds an overlay of the
potential energy surface for the corresponding dataset.
For inject_PES() a colormap is defined by
make_png_with_bad_as_transparent_colormap(). This colormap allows plotting
of PNGs where white pixels are set to transparent.