Adapt correct_2D_grid to save mean values not only under the key (x,y) but also (y,x)
When calculating the corrected_2D array for model predictions the values
are stored under a key, from mich the can be accessed. However, if the key
is not in the dictionary no prediction can be made. A heat map at x,y however
has the same values as a heat map positioned at y,x in a meat grid and is
only transposed.
So far no plotting of these transposed maps as single maps was possible.
Therefore after calculating the mean values in correct_2D_grid, now the
result is stored both under the key (x,y) and (y,x), allowing for the plotting
of such transposed maps.