Add option to inject linear dividing surface into plot_single_map()
Apart from a PES overlay sometimes an already postulated dividing surface
should be added to a plot.
plot_single_map() was therefore extended to potentially take a second injected
function to plot a line graph over the head map. The function to plot the
line graph was already present in NPA and was moved to plotter.py.
Functions to calculate the slope of different postulated dividing surfaces
were moved from NPA to helperfunctions.py
Calls to plot_single_map() for MCGvsBigCage and NoW3vsBigCage were added to NPA
for Groundtruth and Predictions plots.