Add transposing of matrix back into plot_single_map()
Previously the unavailability of mean values for heatmaps above the diagonal
forced a change in plot_single_map() to remove the transposing of the calculated
output and at the same time give x_int as input for y_int and vice versa.
As this is confusing and unclean, and the corrected_2D dict now also contains
keys to the above-diagonal mean values, this change was reverted.