Fix incorrect tick labels for plot_single_map()
Since the order in the pipeline was changed and the removal of unneeded
dimensions happens after the bounding, the tick labels for all heat map
plots have been incorrect.
This was due to the fact, that they accessed the wrong positions in
pipeline.lower_bound and pipeline.upper_bound, as these lists are still
full dmensional, while the int given is for the reduced-dimensional list.
To fix this issue, the indices of the reduced list are now used to look
up the corresponding variable name in reduced_list_var_names. This name
is then used to look up the corresponding position in
const.name_to_list_postion. This new index is then used for all pipeline
lookups and corresponds to the original index on the reduced list.