Change global constants to improve performance
Change the encoder activation function from "linear" to "tanh" to allow for
the capturing of non-linear relationships.
Change the ratio of label vs reconstruction loss weights from 1:1 to 1:0.1, as
the previous ratio sometimes caused the model to make poor predictions.
Change the logvmin from 10**(-10) to 10**(-4) to make visualization of plots
clearer.