Refunction the functions for plotting on latent space to also map paths to their commitor outputs
plot_example_paths_on_latent_space() so far only mapped paths on the latent
space. However, it can also be used to map paths onto the committor space
as that is also one dimensions.
Options for plotting onto a 2D latent space were removed as they are not
needed and only cluttered the code.
Functions and variable were renamed to replace "latent" with "projected".
The ylabel of the plot was changed to be based on the output_name of the
model used.
A call to the renamed function plot_projected_example_paths() was added to NPA.