Change get_mean_values_per_point() to return a dict of dicts and simplify calculations in calc_represented_map_generated() and calc_represented_scatter_generated()
calc_represented_map_generated() and calc_represented_scatter_generated() were
hard to follow due to the fact that the representation was a dict of lists.
By changing the representation to a list of lists, the two functions could
be simplified.
Inner functions resclae_grid_point_means() were defined and the grid_point_means
evaluated together with their respective key (the position on the grid) by
iterarting over the sub-dictionary"